WIP: Added monitor. Moved stuff from uxncli to wioterminal. Before adding screen device
