+++ src/monitor.h
@@ -1,0 +1,5 @@
+/**
+ * A small machine code monitor
+ */
+void mon_init();
+void mon_onechar(Uxn* u, char c);
